Modified watchdog.sh to not kill the container before first visit

+++ watchdog.sh
@@ -1,10 +1,13 @@
 #!/bin/bash
-
+visited=0
 while true; do
 	sleep 60
 	netstat -t | grep ':8004'
-	if [ $? != 0 ]; then
+	ret=$?
+	if [ $ret != 0 -a $visited -eq 1 ]; then
 		kill $(cat /var/run/supervisord.pid)
+	elif [ $ret == 0 -a $visited -eq 0 ]; then
+		visited=1
 	fi
 done
